+++ rpc/backend/call_tx.go
@@ -299,7 +299,8 @@ func (b *Backend) SetTxDefaults(args evmtypes.TransactionArgs) (evmtypes.Transac
 		}
 
 		blockNr := rpctypes.NewBlockNumber(big.NewInt(0))
-		estimated, err := b.EstimateGas(callArgs, &blockNr)
+		blockNrOrHash := rpctypes.BlockNumberOrHash{BlockNumber: &blockNr}
+		estimated, err := b.EstimateGas(callArgs, &blockNrOrHash, nil)
 		if err != nil {
 			return args, err
 		}
@@ -317,11 +318,16 @@ func (b *Backend) SetTxDefaults(args evmtypes.TransactionArgs) (evmtypes.Transac
 // EstimateGas returns an estimate of gas usage for the given smart contract call.
 func (b *Backend) EstimateGas(
 	args evmtypes.TransactionArgs,
-	blockNrOptional *rpctypes.BlockNumber,
+	blockNrOrHash *rpctypes.BlockNumberOrHash,
+	overrides *json.RawMessage,
 ) (hexutil.Uint64, error) {
 	blockNr := rpctypes.EthPendingBlockNumber
-	if blockNrOptional != nil {
-		blockNr = *blockNrOptional
+	if blockNrOrHash != nil {
+		var err error
+		blockNr, err = b.BlockNumberFromComet(*blockNrOrHash)
+		if err != nil {
+			return 0, err
+		}
 	}
 
 	bz, err := json.Marshal(&args)
@@ -335,11 +341,17 @@ func (b *Backend) EstimateGas(
 		return 0, errors.New("header not found")
 	}
 
+	var bzOverrides []byte
+	if overrides != nil {
+		bzOverrides = *overrides
+	}
+
 	req := evmtypes.EthCallRequest{
 		Args:            bz,
 		GasCap:          b.RPCGasCap(),
 		ProposerAddress: sdk.ConsAddress(header.Header.ProposerAddress),
 		ChainId:         b.EvmChainID.Int64(),
+		Overrides:       bzOverrides,
 	}
 
 	// From ContextWithHeight: if the provided height is 0,