+++ CHANGELOG.md
@@ -78,7 +78,7 @@
 - [\#710](https://github.com/cosmos/evm/pull/710) Fix EoA-CA Identification logic
 - [\#711](https://github.com/cosmos/evm/pull/711) Add debug_traceCall api
 - [\#734](https://github.com/cosmos/evm/pull/734) Disable evm mempool if max-txs set to -1.
-
+- [\#743](https://github.com/cosmos/evm/pull/743) Apply state overrides to eth_estimateGas api
 
 ### FEATURES